Call the warframe api from one function
Moved the const declarations for the api to the start of the file and
added a new function that calls the warframe api and return the api
promise.
This way, making calls to the warframe api only requires the "noun"
(voidTrader, alerts etc.) and we don't need to redaclare the constant on
every api request.